--- Client/Frontend/Widgets/PhotonActionSheet/AppMenu.swift
@@ -1,105 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0
-
-import Shared
-import Account
-
-extension PhotonActionSheetProtocol {
-
-    func getSettingsAction(vcDelegate: MenuActionsDelegate?) -> PhotonActionSheetItem {
-        // This method is being called when we the user sees the menu, not just when it's constructed.
-        // In that case, we can let sendExposureEvent default to true.
-        let variables = Experiments.shared.getVariables(featureId: .nimbusValidation)
-        // Get the title and icon for this feature from nimbus.
-        // We need to provide defaults if Nimbus doesn't provide them.
-        let title = variables.getText("settings-title") ?? .AppMenuSettingsTitleString
-        let icon = variables.getString("settings-icon") ?? "menu-Settings"
-
-        let openSettings = PhotonActionSheetItem(title: title, iconString: icon) { _, _ in
-            let settingsTableViewController = AppSettingsTableViewController()
-            settingsTableViewController.profile = self.profile
-            settingsTableViewController.tabManager = self.tabManager
-            settingsTableViewController.settingsDelegate = vcDelegate
-            
-            let controller = ThemedNavigationController(rootViewController: settingsTableViewController)
-            // On iPhone iOS13 the WKWebview crashes while presenting file picker if its not full screen. Ref #6232
-            if UIDevice.current.userInterfaceIdiom == .phone {
-                controller.modalPresentationStyle = .fullScreen
-            }
-            controller.presentingModalViewControllerDelegate = vcDelegate
-            TelemetryWrapper.recordEvent(category: .action, method: .open, object: .settings)
-            
-            // Wait to present VC in an async dispatch queue to prevent a case where dismissal
-            // of this popover on iPad seems to block the presentation of the modal VC.
-            DispatchQueue.main.async {
-                // TODO: Laurie - use other delegate? standardize
-                vcDelegate?.present(controller, animated: true, completion: nil)
-            }
-        }
-        return openSettings
-    }
-    
-    func getNightModeAction() -> [PhotonActionSheetItem] {
-        var items: [PhotonActionSheetItem] = []
-
-        let nightModeEnabled = NightModeHelper.isActivated(profile.prefs)
-        let nightModeTitle: String = nightModeEnabled ? .AppMenuTurnOffNightMode : .AppMenuTurnOnNightMode
-        let nightMode = PhotonActionSheetItem(title: nightModeTitle, iconString: "menu-NightMode", isEnabled: nightModeEnabled) { _, _ in
-            NightModeHelper.toggle(self.profile.prefs, tabManager: self.tabManager)
-
-            if nightModeEnabled {
-                TelemetryWrapper.recordEvent(category: .action, method: .tap, object: .nightModeEnabled)
-            } else {
-                TelemetryWrapper.recordEvent(category: .action, method: .tap, object: .nightModeDisabled)
-            }
-
-            // If we've enabled night mode and the theme is normal, enable dark theme
-            if NightModeHelper.isActivated(self.profile.prefs), LegacyThemeManager.instance.currentName == .normal {
-                LegacyThemeManager.instance.current = DarkTheme()
-                NightModeHelper.setEnabledDarkTheme(self.profile.prefs, darkTheme: true)
-            }
-            // If we've disabled night mode and dark theme was activated by it then disable dark theme
-            if !NightModeHelper.isActivated(self.profile.prefs), NightModeHelper.hasEnabledDarkTheme(self.profile.prefs), LegacyThemeManager.instance.currentName == .dark {
-                LegacyThemeManager.instance.current = NormalTheme()
-                NightModeHelper.setEnabledDarkTheme(self.profile.prefs, darkTheme: false)
-            }
-        }
-        items.append(nightMode)
-
-        return items
-    }
-
-    func syncMenuButton(showFxA: @escaping (FXAClosureType) -> Void) -> PhotonActionSheetItem? {
-        let action: ((PhotonActionSheetItem, UITableViewCell) -> Void) = { action,_ in
-            let fxaParams = FxALaunchParams(query: ["entrypoint": "browsermenu"])
-            let params = FXAClosureType(fxaParams, .emailLoginFlow, .appMenu)
-            showFxA(params)
-            TelemetryWrapper.recordEvent(category: .action, method: .tap, object: .signIntoSync)
-        }
-
-        let rustAccount = RustFirefoxAccounts.shared
-        let needsReauth = rustAccount.accountNeedsReauth()
-
-        guard let userProfile = rustAccount.userProfile else {
-            return PhotonActionSheetItem(title: .AppMenuBackUpAndSyncData, iconString: "menu-sync", handler: action)
-        }
-        let title: String = {
-            if rustAccount.accountNeedsReauth() {
-                return .FxAAccountVerifyPassword
-            }
-            return userProfile.displayName ?? userProfile.email
-        }()
-
-        let iconString = needsReauth ? "menu-warning" : "placeholder-avatar"
-
-        var iconURL: URL? = nil
-        if let str = rustAccount.userProfile?.avatarUrl, let url = URL(string: str) {
-            iconURL = url
-        }
-        let iconType: PhotonActionSheetIconType = needsReauth ? .Image : .URL
-        let iconTint: UIColor? = needsReauth ? UIColor.Photon.Yellow60 : nil
-        let syncOption = PhotonActionSheetItem(title: title, iconString: iconString, iconURL: iconURL, iconType: iconType, iconTint: iconTint, handler: action)
-        return syncOption
-    }
-}
